requirements.txt
# Web scraping and browser automation
selenium>=4.0.0
playwright-stealth>=1.0.6
webdriver-manager>=4.0.0

# Google API integration
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# HTTP requests
requests>=2.31.0
aiohttp>=3.8.1

# Environment variables
python-dotenv>=1.0.0

# JSON handling (included in Python standard library, but listing for clarity)
# json - built-in